backend/requirements.txt
anyio==4.3.0
click==8.1.7
colorama==0.4.6
exceptiongroup==1.2.1
fastapi==0.95.1
greenlet==3.0.3
h11==0.14.0
idna==3.7
pydantic==1.10.15
python-dotenv==1.0.1
sniffio==1.3.1
SQLAlchemy==2.0.30
starlette==0.26.1
typing_extensions==4.11.0
uvicorn==0.22.0
